# openapi.yaml
openapi: 3.0.3
info:
  title: RASS API
  description: Reference Architecture for Semantic Search - API specification for document embedding and semantic search
  version: 1.0.0
  contact:
    name: RASS API Support
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    EmbeddedDocument:
      type: object
      required:
        - id
        - content
        - metadata
        - created_at
        - updated_at
      properties:
        id:
          type: string
        content:
          type: string
        metadata:
          type: object
          additionalProperties: true
        embedding:
          type: array
          items:
            type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
security:
  - ApiKeyAuth: []
paths:
...
  /item/{id}:
    get:
      summary: Get a document by ID
      description: Retrieve a specific embedded document by its ID
      tags:
        - Documents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Document ID
      responses:
        '200':
          description: Document found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbeddedDocument'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...